+++ change-log.md
@@ -3,6 +3,8 @@
 
 ## Genshin Impact v. 6.6
 
+## New Features
+* Profile avatar: You can now change your profile avatar to whichever character you want, instead of being just limited to Lumine and Aether!
 
 ## Updates
 * Added "Off-field Damage" to Lauma